refactor: introduce application layer in Command
