+++ src/php/Command/CommandFactory.php
@@ -6,19 +6,19 @@ namespace Phel\Command;
 
 use Gacela\Framework\AbstractFactory;
 use Gacela\Framework\Config\ConfigReader\PhpConfigReader;
+use Phel\Command\Application\CommandExceptionWriter;
+use Phel\Command\Application\DirectoryFinder;
+use Phel\Command\Application\TextExceptionPrinter;
 use Phel\Command\Domain\Finder\ComposerVendorDirectoriesFinder;
-use Phel\Command\Domain\Finder\DirectoryFinder;
 use Phel\Command\Domain\Finder\DirectoryFinderInterface;
 use Phel\Command\Domain\Finder\VendorDirectoriesFinderInterface;
-use Phel\Command\Domain\Handler\ExceptionHandler;
-use Phel\Command\Domain\Shared\CommandExceptionWriter;
 use Phel\Command\Domain\Shared\CommandExceptionWriterInterface;
 use Phel\Command\Domain\Shared\ErrorLog\ErrorLog;
 use Phel\Command\Domain\Shared\Exceptions\ExceptionArgsPrinter;
 use Phel\Command\Domain\Shared\Exceptions\ExceptionPrinterInterface;
 use Phel\Command\Domain\Shared\Exceptions\Extractor\FilePositionExtractor;
 use Phel\Command\Domain\Shared\Exceptions\Extractor\SourceMapExtractor;
-use Phel\Command\Domain\Shared\Exceptions\TextExceptionPrinter;
+use Phel\Command\Infrastructure\ExceptionHandler;
 use Phel\Compiler\Application\Munge;
 use Phel\Printer\Printer;
 use Phel\Run\Domain\Repl\ColorStyle;
